backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
SQLAlchemy==2.0.30
psycopg[binary]==3.1.19
boto3==1.34.131
python-jose[cryptography]==3.3.0
httpx==0.27.0
slowapi==0.1.9
